Silme İşlemleri Tamamladı
Silme İşlemleri tamamlandı.
Sayfalarda şablonlar oluşturuldu.
Buglar düzeltildi.

+++ src/main/java/org/freesoftware/dao/PersonDAO.java
@@ -9,7 +9,7 @@ public interface PersonDAO {
 
 	public List<Person> getAllPersons();
 
-	public void deletePerson(Integer personId);
+	public void deletePerson(long personId);
 
 	public Person updatePerson(Person person);
